A quick fix for hypervocal's border on the body el.  This is needed because jQuery's offset function doesn't take the body's border into account.  Later, we'll want to be a little more thorough with this fix, but it's fine for now- it fixes the problem of making the indicators look right on hypervocal.
Example of what still should be considered in this fix: - the stay-in-window function doesn't compensate for the body border... but it doens't matter for a small border anyway, so we're good for now.  It should be a rare case that a page would have a big border on the body el, but you never know I guess.