fixing tumblr - it was looking for kind: "txt" but the value was "text" - checking for both now.
It's all mostly working - the image needs to be fixed though - it doesn't send the image right.  I think the bookmarklet was updated and now it needs a new format.  I couldn't figure out for the life of me how to send the right thing…  Will look again later.

Also note that the tumblr media assumes the video was in an iframe (will work for youtube).  Take others into account later.